% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionitio.R
\name{get_seq_stats}
\alias{get_seq_stats}
\title{Calculate sequence statistics}
\usage{
get_seq_stats(sequences, filename = NULL, min_len = 0)
}
\arguments{
\item{sequences}{A character vector containing sequences}

\item{filename}{The filename of the FASTA file}

\item{min_len}{The minimum length of the sequence to include when
calculating statistics}
}
\value{
A list containing FASTA stats.
}
\description{
Calculate the number of sequences (num_seq), the length of
    the smallest sequence (min_seq), the length of the largest sequence
    (max_seq), and the total number of bases (num_bases) in the vector of
    given sequences. A minimum length threshold can be set to exclude
    sequences less than the specified size.
}
\examples{
my_fasta_file <- system.file("extdata", "two_sequence.fasta", package = "bionitio")
sequences <- load_fasta_file(filename = my_fasta_file)
get_seq_stats(sequences, filename = my_fasta_file)
}

