% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionitio.R
\name{load_fasta_file}
\alias{load_fasta_file}
\title{Load a FASTA file}
\usage{
load_fasta_file(filename, quit_on_error = TRUE)
}
\arguments{
\item{filename}{The name of the input FASTA file.}

\item{quit_on_error}{A logical value indicating if an error when reading a
fasta file will invoke the \code{quit} function to terminate R with an
exit status of 3. This will only occur if the R session is
non-interactive.}
}
\value{
A character vector of containing sequenecs as characters. Each
    element in the vector is a sequence in the FASTA file.
}
\description{
Wrapper for seqinr \code{read.fasta()} to load in a FASTA file
    and get the sequences.
}
\examples{
my_fasta_file <- system.file("extdata", "two_sequence.fasta", package = "bionitio")
load_fasta_file(filename = my_fasta_file)
}

